# OperationsApi.Tests/bin/Debug/net10.0/openapi/operations-api.yaml
openapi: 3.1.0
info:
  title: Dashboard API
  version: 1.0.0
  description: |
    Contract for the wind-farm dashboard. The API reads the provided telemetry
    CSV files and owns validation, aggregation, alert evaluation, and data-quality
    detection. All summaries and series are derived solely from telemetry rows;
    the service does not manufacture readings for farms without telemetry.
servers:
  - url: /api/operations/
    description: Dashboard API base path
tags:
  - name: Dashboard
  - name: Farms
  - name: Turbines
  - name: Alerts
paths:
  /dashboard:
    get:
      tags: [Dashboard]
      summary: Get fleet dashboard metrics
      operationId: getDashboard
      responses:
        '200':
          description: Fleet metrics and telemetry-derived farm summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
  /farms/{farmId}:
    get:
      tags: [Farms]
      summary: Get a farm and its turbines
      operationId: getFarm
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: The farm and its turbines.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'
        '404':
          $ref: '#/components/responses/NotFound'
  /turbines/{turbineId}:
    get:
      tags: [Turbines]
      summary: Get a turbine summary and telemetry series
      operationId: getTurbine
      parameters:
        - $ref: '#/components/parameters/TurbineId'
      responses:
        '200':
          description: The turbine summary and its available telemetry series.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turbine'
        '404':
          $ref: '#/components/responses/NotFound'
  /alerts:
    get:
      tags: [Alerts]
      summary: List alert presentation records
      operationId: listAlerts
      responses:
        '200':
          description: Alerts evaluated from the provided telemetry data.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
components:
  parameters:
    FarmId:
      name: farmId
      in: path
      required: true
      description: Unique farm identifier.
      schema:
        type: string
    TurbineId:
      name: turbineId
      in: path
      required: true
      description: Unique turbine identifier.
      schema:
        type: string
  responses:
    NotFound:
      description: The requested farm or turbine was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Dashboard:
      type: object
      required: [fleetMetrics, farms]
      properties:
        fleetMetrics:
          $ref: '#/components/schemas/FleetMetrics'
        farms:
          type: array
          description: Farm summaries calculated only where telemetry is available.
          items:
            $ref: '#/components/schemas/FarmDashboardSummary'
    FleetMetrics:
      type: object
      description: Aggregate values calculated from the provided telemetry rows.
      properties:
        averagePowerKw:
          type: number
          format: double
        averageWindMs:
          type: number
          format: double
        criticalAlertCount:
          type: integer
          minimum: 0
    FarmDashboardSummary:
      type: object
      required: [farmId, farmName, averagePowerKw, averageWindMs, criticalAlertCount]
      properties:
        farmId:
          type: string
        farmName:
          type: string
        averagePowerKw:
          type: number
          format: double
        averageWindMs:
          type: number
          format: double
        criticalAlertCount:
          type: integer
          minimum: 0
    Farm:
      type: object
      required: [farmId, farmName, turbines]
      properties:
        farmId:
          type: string
        farmName:
          type: string
        turbines:
          type: array
          items:
            $ref: '#/components/schemas/FarmTurbine'
    FarmTurbine:
      type: object
      required: [turbineId]
      properties:
        turbineId:
          type: string
    Turbine:
      type: object
      required: [turbineId, farmId, farmName, powerOutput, windSpeed]
      properties:
        turbineId:
          type: string
        turbineName:
          type: string
        farmId:
          type: string
        farmName:
          type: string
        averagePowerKw:
          type: number
          format: double
        averageWindMs:
          type: number
          format: double
        criticalAlertCount:
          type: integer
          minimum: 0
        powerOutput:
          $ref: '#/components/schemas/TelemetrySeries'
        windSpeed:
          $ref: '#/components/schemas/TelemetrySeries'
    TelemetrySeries:
      type: object
      required: [unit, points]
      properties:
        unit:
          type: string
          description: Unit for every point in this series, such as kW or m/s.
        points:
          type: array
          items:
            $ref: '#/components/schemas/TelemetryPoint'
    TelemetryPoint:
      type: object
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
          format: double
    Alert:
      type: object
      required: [category, severity, timestamp, title, explanation]
      properties:
        category:
          type: string
          description: Alert classification, including data-quality categories where applicable.
        severity:
          type: string
          enum: [info, warning, critical]
        farmId:
          type: string
        farmName:
          type: string
        turbineId:
          type: string
        turbineName:
          type: string
        timestamp:
          type: string
          format: date-time
        title:
          type: string
        explanation:
          type: string
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string